Added Amount Limit [SPEN-19]
Added condition to submitEntry that forcibly sets amount to "$ 9,999,999.99" if the input is ten million or more, which is checked by analyzing the length of the input. Fourteen characters would be less than ten million, but fifteen characters would be ten million or more, which includes a dollar sign, a space, two commas, and a decimal point.
Also, added the replace method to the amount initialization in addEntry in DataManager, which will remove any commas from the amount.