Added Date Validation [SPEN-19]
Added conditionals to submitEntry that confirm that date input are valid format. If date is not valid format, then function will add the entry anyway, but will forcibly set the date to the current date, thus completely ignoring the user's invalid input. The user can change the date later in Tab 3.