Implemented Category Archiving [SPEN-38]
Performed a successful test of changing a category.
The category was entirely saved into the Archive.xml.
The category's year children were all removed from the WalletWatch.xml.
The category label was correctly modified in the WalletWatch.xml.
The total of all data was accurately decremented by the amount of the old category in the WalletWatch.xml.
Lastly, the category amount was reset to 0.00 in the WalletWatch.xml.
The function changeCategories correctly only changed categories that were indicated as changed and ignored null values in the labels parameter.
All around a successful test.