Modified MIT License [SPEN-65]
Changed organization name from "Team 4" to "SpendSages".
Placed MIT License in content of Terms tab being that the WalletWatch app is a free, open-source and offline mobile app. No ToS, EULA, etc. are necessary or even make sense since SpendSages is not a company and WalletWatch is not a commercial product.