Removed Whitespace from XML Files [SPEN-57]
Whitespace in the XML files were being parsed as their own Element/Node/Tags.
Also fixed a bug related to accessing the Category 1 textbox. This bug caused the "master" branch code to crash, which is why this commit is being pushed directly to master rather than in a branch SPEN-57.