True Bugfix of Blank Total [SPEN-55]
Setting focus to a drop-down menu did not actually fix the bug of blank total(s) on Tab 2. Instead, the problem actually existed in the constraints setup in the UI of fragment_tab2.xml. Using 'wrap_content" instead of a set constraint resulted in the total sometimes wrapping to 0 pixels, for some unknown reason. It should now be fixed, for reals this time.