Fixed Sorting/Filtering Bug [SPEN-54]
Deleting expenses on Tab 3 would then resort and unfilter the list. Now the sorting and filtering match the current positions of the Spinners.
Also changed label to "Submit the Expense by tapping a Category:" to make it more clear to first-time users that tapping a category button will immediately submit the entry.