Resolved Recycler Refresh Build Warning
The notifyDataSetChanged method is considered inefficient, since it will refresh the entire list in the UI. Instead, we can utilize a helper function to identify the exact position of the entry and then notify the UI that only that specific entry/index needs to be redrawn. The slight exception is the sorting function, which is meant to redraw the entire list. Nonetheless, a range method was utilized to avoid the build warning.