Resolved onBackPressed Build Warning
The onBackPressed method was deprecated in favor of a few properties within the "activity" element inside AndroidManifest.XML, such as android:noHistory and android:exported as well as the intent-filter element. This changeset has the added benefit of persisting user inputs on the MainActivity after visiting the SettingsActivity. However, user inputs on the SettingsActivity are not persisted, which is acceptable since category edits are expected to be an infrequent user event.