Implemented Data Repository
Live data is now working again. Due to the switch to ViewPager2 and TabLayoutMediator, the entire live data refresh paradigm had to be overhauled. However, live data is not yet successfully transferring from the SettingsActivity to the MainActivity, namely renaming a category does not auto-refresh the Tab 1 category button name.