Fixed Category Restore Bug
Since child category nodes were being moved out of the parent node and adopted into a new parent node, the category NodeList was getting smaller with fewer and fewer Node elements, but the for loop conditional was still checking the original NodeList size. This resulted in IndexOutOfBounds exceptions. The solution was to create a deep copy list of the nodes and iterate through those clones, but only move/adopt the original non-clone child nodes.