Added Backup and Restore Feature
Since the adb backup and restore commands are deprecated and soon to be removed, we have added a feature that allows us as developers and our users the ability to backup/export and restore/import data. To avoid requiring our users to accept file system permissions, the backup feature simply copies the current content of WalletWatch.XML or Archive.XML to the user's clipboard. Similarly, the restore feature is implemented by an editable textbox with basic, minimal input validation.