Fixed Space Key Typing Bug
Moved trimming of leading and trailing whitespace to the corresponding save/submit function, rather than forcibly overwriting the content in the textbox inside an onTextChangedListener. This bug was preventing users from tapping the spacebar unless the cursor was in the middle of the string. Also optimized the checkInputs function in CategoryFragment.kt.