Added Deselect All Entries to Tab3Fragment
Filled in the scaffolding needed to support the new feature that deselects all entries while also providing the number of selected entries and their total dollar sum. This includes the addition of a selectedSum variable, a variable for the checkbox itself, a variable for the Toast message, visibility toggling logic, checkbox state logic, OnClickListener logic, Toast message logic, and reset of the checkbox text and selectedSum to zero values. Also replaced lines for the deleteButton visibility logic with a simple call of the toggleButton helper function. Placed TODO comments with implementation reminders to be done in a follow-up commit.