Added New setEntryCheckBox Method in RecyclerAdapter
The purpose of the new setEntryCheckBox method was to provide Tab3Fragment an easy way to set the state of an Entry's selected bool member variable. The unfiltered list entriesRaw was used because the user may have selected an entry from one category, but then changed the filter dropdown to a different category. So, we want to deselect the entry's checkbox even though it is currently not visible.