Fixed Tab2Fragment Frozen/Read-only Bug
The detach and attach of the same fragment transaction is required in our use case; otherwise, Tab2Fragment freezes and the user is not able to tap any buttons or navigate to a different tab. By consolidating back down to the detach and attach, along with a new Boolean to track whether or not to perform the re-attach/refresh, the bug was fixed. The warning regarding the possible no-op has been suppressed, especially since we are calling the setReorderingAllowed(false) method that explicitly prevents the fragment manager from performing a no-op of re-attachments.