Moved Metric Prefix Helper Method in Tab2Fragment to MainActivity
Moved the helper function formatTotal from Tab2Fragment up to the MainActivity and renamed it to formatDollarAmount. Also moved the counterpart array constant "metric" to the MainActivity as well. By moving these up to the activity level and making them public, the Tab3Fragment is now able to utilize the metric prefix string formatting power of this helper method, which explains the changes that greatly simplified the updateDeselectAllCheckBoxText helper method implementation inside Tab3Fragment.