Fixed Line Chart Y-axis Label Bug
A bug was occurring whenever there was a Y-axis value that required one or more labels to utilize a metric prefix. The root cause was calcualting the yValExponent with the Float yValue. By converting it to a Double before passing it into the "log10" math function, this eliminated the lack of precision that was occurring. Subsequently, other hardcoded comparison values in the if conditionals below had to be switched from Float to Double as well. Also made the code more readable by moving the metric prefix character retrieval logic into local variables.