Fixed Invalid Amount Red 'X' Display State Logic Bug
If the user forgot to enter in a valid amount and skipped ahead to entering in either the description or date, then they would not realize that their amount input being left at zero "$ 0.00" is invalid. By fixing this logic to ensure that the invalid red 'X' does properly show in all situations where it ought to, the user experience is improved, since they will now know why the category buttons are still disabled.