Fixed Entry Edit Save Changes Button State Bug
Under certain conditions, the Save Changes button on the entry edit window would be enabled when it really ought to have been disabled due to an invalid user input. By explicitly checking the amount and date textboxes for invalid input, we can bypass the check for changes to the entry data and forcibly disable the Save Changes button.