Added Tab Swipe Navigation Progress Bar Feature
Once users eventually accumulated a large number of expense entries (500+), they started to notice that swiping to the History Tab3Fragment was slow on the first try of each app session. The root cause of the slowdown appears to be population of the RecyclerAdapter EntryViewHolder elements. The model.get() function calls seem to be very fast still, but actually populating the full expense history list UI cards is the bottleneck. As a benchmark, 707 entries caused a UI lockup (frozen screen) for about 5 seconds on a Debug build on a Google Pixel 5a. A prior commit addressed some of the redundant operations, which helped boost the performance by cutting the time down to about 3 seconds. The changes in this commit implement a round indeterminate progress bar circle that briefly appears when the user swipes to another tab in the MainActivity. Once the user navigates to the History Tab3Fragment, the progress bar permanently disappears (until the app session is closed and a new app session is launched later), since the tab swipe navigation slowdowns completely disappear. Unfortunately, we have not determined how to show the progress bar when the user directly tabs Tab 2 or Tab 3, as opposed to swiping.