Bugfixes and Improvements
Strategically moved code from Tab3Fragment into RecyclerAdapter. Fixed a bug where changing the category of a selected entry was unchecking the checkbox after saving the changes. This bugfix required returning the entry ID from both the add and edit methods in the DataManager. Also added logic to only refresh the category labels on the MainActivity if a category label was actually changed in the SettingsActivity.