Implemented DiffUtil
Re-architected the Recycler Adapter that displays the list of entries on the Tab 3 Fragment to the DiffUtil framework, which provides responsive animations and massive performance improvements. This included defining an EntryDiffCallback class, moving the sorting algorithms out of the Entry class and into the RecyclerAdapter class, overhauling the "controller" methods of the RecyclerAdapter and Tab3Fragment (initialize, filter, sort, delete, edit, select, deselect), and added function comments.