Fixed Tab 3 Category Refresh Bug
The submitCategories function in the RecyclerAdapter was not at all what needed to happen when the user changed the category labels. Instead of re-assigning the category label of the existing entries in the recycler view, we need to forcibly re-initialize the full list of entries. This ensures that archived category entries are removed from the recycler and restored/new category entries are added to the recycler.