Fixed ViewPager2 Transition Bugs
If the user launched the app and tapped Tab 3, but then quickly tapped empty space on Tab 1, the app would appear to freeze. By disabling user input while the tab selection ViewPager2 swipe animation is occurring, we avoid the issue of conflicting user inputs regarding the selected tab fragment. Additionally, made some fixes to the logic regarding the hiding of the keyboard/numpad.